--- a/fedimint-server/src/lib.rs
+++ b/fedimint-server/src/lib.rs
@@ -14,6 +14,9 @@
 #![allow(clippy::similar_names)]
 #![allow(clippy::too_many_lines)]
 #![allow(clippy::needless_pass_by_value)]
+#![allow(clippy::manual_let_else)]
+#![allow(clippy::match_wildcard_for_single_variants)]
+#![allow(clippy::trivially_copy_pass_by_ref)]
 
 extern crate fedimint_core;