This pr is part of a larger effort to simplify the p2p networking situation.

Removing the parallelization of the DKG across modules and across keys allows us to completely remove the multiplexing wrapper around the p2p networking layer and allows for further simplifications in  the server. As a downside the DKG is about 45x slower, primarily driven by the mint keys. Generating one keyset should take about 4x one way latency between peers hence about 200ms - 400ms or 4.5s - 9s for the entire DKG in a real deployment while it remains negligible in testing. 

We now explicitly rely on the assumption that the networking layer is reliable during the dkg, meaning all messages are delivered in order. The current design appears to try to work on an unreliable layer as well but fails to do so anyway as it's fundamentally impossible so we might as well exploit the reliability assumption for a cleaner design.

Before the DKG we wait for all p2p connections to be open and after the DKG we now await our outgoing message queues to drain in order for our fellow peers to complete the DKG as well; this should make the DKG more reliable especially with a high number of peers.




   <!--

# Code Review Policy

* CI must pass (enforced)
* 1 review is mandatory (enforced), 2 or more ideal
* If you believe your change is simple, and non-controversial enough, and you want
  to avoid merge conflicts, or blocking work before it gets enough reviews, label it with
  `needs further review` label and Merge it.

See https://github.com/fedimint/fedimint/blob/master/CONTRIBUTING.md#code-review-policy for
full description.

-->